refactor: moves referral logic to a hook
